move creation of Cart data to server
