+++ utils/get-cart-total.ts
@@ -1,0 +1,7 @@
+import { CartItem } from '../types'
+
+export const getCartTotal = (items: CartItem[]): number =>
+  items.reduce(
+    (total, item) => total + (item.count * item.price - item.discount),
+    0
+  )
